Change folder structure for easier plotting
