+++ visualize_events/dataloader.py
@@ -1,7 +1,6 @@
 # from ..utils.measurements import get_folder_path
 import os
 import json
-from datetime import datetime
 from dateutil import parser, relativedelta
