+++ visualize_events/figures/local_run_1.png
