+++ visualize_events/figures/local_run_2.png
