+++ visualize_events/figures/local_run_3.png
