+++ visualize_events/figures/network_run2_err.png
