+++ visualize_events/figures/network_run4_err.png
