+++ visualize_events/figures/network_run5_err.png
